# contracts/openapi/v1/secondbox.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "SecondBox API",
    "version": "1.0.0",
    "description": "Provider-neutral administration and durable Sandbox API."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "serviceAccountBearer": []
    }
  ],
...
  "paths": {
...
    "/v1/sandboxes/{sandboxId}/terminals": {
      "post": {
        "tags": ["Execution"],
        "operationId": "createSandboxTerminal",
        "parameters": [
          {"$ref": "#/components/parameters/SandboxID"},
          {"$ref": "#/components/parameters/Generation"},
          {"$ref": "#/components/parameters/LeaseIDHeader"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateTerminalRequest"}}}
        },
        "responses": {
          "201": {
            "description": "Reconnectable PTY session.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TerminalSession"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "429": {"$ref": "#/components/responses/TooManyRequests"}
        }
      }
    },
    "/v1/sandboxes/{sandboxId}/terminals/{terminalSessionId}": {
      "get": {
        "tags": ["Execution"],
        "operationId": "reconnectSandboxTerminal",
        "description": "Returns an authenticated WebSocket connection descriptor for a detached PTY.",
        "parameters": [
          {"$ref": "#/components/parameters/SandboxID"},
          {"$ref": "#/components/parameters/TerminalSessionID"},
          {"$ref": "#/components/parameters/Generation"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "responses": {
          "200": {
            "description": "PTY session connection descriptor.",
            "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TerminalSession"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"}
        }
      },
      "delete": {
        "tags": ["Execution"],
        "operationId": "cancelSandboxTerminal",
        "parameters": [
          {"$ref": "#/components/parameters/SandboxID"},
          {"$ref": "#/components/parameters/TerminalSessionID"},
          {"$ref": "#/components/parameters/Generation"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/RequestID"}
        ],
        "responses": {
          "202": {"$ref": "#/components/responses/AcceptedOperation"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceAccountBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SecondBox service-account API key"
      }
    },
    "headers": {
      "RequestID": {
        "description": "End-to-end request correlation identifier.",
        "schema": {"$ref": "#/components/schemas/CorrelationID"}
      },
...
      "IdempotencyReplayed": {
        "description": "True when the original result is replayed.",
        "schema": {"type": "boolean"}
      },
...
    },
    "parameters": {
      "RequestID": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "schema": {"$ref": "#/components/schemas/CorrelationID"}
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {"type": "string", "minLength": 8, "maxLength": 200, "pattern": "^[A-Za-z0-9._~:+/=-]+$"}
      },
...
      "Generation": {
        "name": "SecondBox-Generation",
        "in": "header",
        "required": true,
        "schema": {"type": "integer", "format": "int64", "minimum": 1}
      },
      "LeaseIDHeader": {
        "name": "SecondBox-Lease-ID",
        "in": "header",
        "required": false,
        "schema": {"$ref": "#/components/schemas/OpaqueID"}
      },
...
      "SandboxID": {"name": "sandboxId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/OpaqueID"}},
...
      "TerminalSessionID": {"name": "terminalSessionId", "in": "path", "required": true, "schema": {"$ref": "#/components/schemas/OpaqueID"}},
...
    },
...
    "responses": {
      "AcceptedOperation": {
        "description": "Mutation accepted and observable as an Operation.",
        "headers": {
          "X-Request-ID": {"$ref": "#/components/headers/RequestID"},
          "Idempotency-Replayed": {"$ref": "#/components/headers/IdempotencyReplayed"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Operation"}}}
      },
      "BadRequest": {"$ref": "#/components/responses/Problem400"},
      "Unauthorized": {"$ref": "#/components/responses/Problem401"},
      "Forbidden": {"$ref": "#/components/responses/Problem403"},
      "NotFound": {"$ref": "#/components/responses/Problem404"},
...
      "Conflict": {"$ref": "#/components/responses/Problem409"},
...
      "TooManyRequests": {"$ref": "#/components/responses/Problem429"},
      "Problem400": {
        "description": "Invalid request.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
      "Problem401": {
        "description": "Authentication failed.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
      "Problem403": {
        "description": "Authorization failed.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
      "Problem404": {
        "description": "Resource not found in the authenticated scope.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
...
      "Problem409": {
        "description": "Current state, generation, Lease, quota, or idempotency conflict.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      },
...
      "Problem429": {
        "description": "Quota or concurrent-operation limit exceeded.",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestID"}},
        "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}
      }
    },
    "schemas": {
      "OpaqueID": {"type": "string", "minLength": 8, "maxLength": 128, "pattern": "^[A-Za-z0-9_-]+$"},
      "CorrelationID": {"type": "string", "minLength": 8, "maxLength": 128, "pattern": "^[A-Za-z0-9._:-]+$"},
      "Timestamp": {"type": "string", "format": "date-time"},
      "ProfileName": {"type": "string", "minLength": 1, "maxLength": 80, "pattern": "^[a-z][a-z0-9-]*$"},
...
      "WorkspacePath": {"type": "string", "minLength": 1, "maxLength": 4096, "pattern": "^(?!/)(?!.*(?:^|/)\\.\\.(?:/|$)).+$"},
      "Metadata": {
        "type": "object",
        "maxProperties": 32,
        "propertyNames": {"type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9._/-]+$"},
        "additionalProperties": {"type": "string", "maxLength": 1024}
      },
      "StringMap": {
        "type": "object",
        "maxProperties": 128,
        "propertyNames": {"type": "string", "minLength": 1, "maxLength": 256},
        "additionalProperties": {"type": "string", "maxLength": 8192}
      },
...
      "SandboxState": {
        "type": "string",
        "enum": ["creating", "stopped", "starting", "ready", "draining", "stopping", "checkpointing", "failed", "deleting", "deleted"]
      },
      "SandboxDesiredState": {"type": "string", "enum": ["running", "stopped", "deleted"]},
      "Sandbox": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "projectId", "profile", "profileRevisionId", "state", "desiredState", "generation", "workspace", "metadata", "revision", "createdAt", "updatedAt"],
        "properties": {
          "id": {"$ref": "#/components/schemas/OpaqueID"},
          "projectId": {"$ref": "#/components/schemas/OpaqueID"},
          "profile": {"$ref": "#/components/schemas/ProfileName"},
          "profileRevisionId": {"$ref": "#/components/schemas/OpaqueID"},
          "state": {"$ref": "#/components/schemas/SandboxState"},
          "desiredState": {"$ref": "#/components/schemas/SandboxDesiredState"},
          "generation": {"type": "integer", "format": "int64", "minimum": 1},
          "workspace": {"$ref": "#/components/schemas/Workspace"},
          "instance": {"$ref": "#/components/schemas/Instance"},
          "metadata": {"$ref": "#/components/schemas/Metadata"},
          "lastActivityAt": {"$ref": "#/components/schemas/Timestamp"},
          "revision": {"type": "integer", "format": "int64", "minimum": 1},
          "createdAt": {"$ref": "#/components/schemas/Timestamp"},
          "updatedAt": {"$ref": "#/components/schemas/Timestamp"},
          "deletedAt": {"$ref": "#/components/schemas/Timestamp"}
        }
      },
...
      "Workspace": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "generation", "retainedBytes", "createdAt", "updatedAt"],
        "properties": {
          "id": {"$ref": "#/components/schemas/OpaqueID"},
          "generation": {"type": "integer", "format": "int64", "minimum": 1},
          "retainedBytes": {"type": "integer", "format": "int64", "minimum": 0},
          "currentSnapshotId": {"$ref": "#/components/schemas/OpaqueID"},
          "retainUntil": {"$ref": "#/components/schemas/Timestamp"},
          "createdAt": {"$ref": "#/components/schemas/Timestamp"},
          "updatedAt": {"$ref": "#/components/schemas/Timestamp"}
        }
      },
      "InstanceState": {"type": "string", "enum": ["starting", "ready", "draining", "stopping", "stopped", "lost", "failed"]},
      "InstanceTerminationReason": {
        "type": "string",
        "enum": ["requested_drain", "requested_stop", "idle_timeout", "maximum_duration", "guest_shutdown", "resource_exhaustion", "guest_agent_lost", "runner_lost", "startup_failed", "fenced", "internal_failure"]
      },
      "Instance": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "sandboxId", "generation", "state", "createdAt", "updatedAt"],
        "properties": {
          "id": {"$ref": "#/components/schemas/OpaqueID"},
          "sandboxId": {"$ref": "#/components/schemas/OpaqueID"},
          "generation": {"type": "integer", "format": "int64", "minimum": 1},
          "state": {"$ref": "#/components/schemas/InstanceState"},
          "terminationReason": {"$ref": "#/components/schemas/InstanceTerminationReason"},
          "createdAt": {"$ref": "#/components/schemas/Timestamp"},
          "readyAt": {"$ref": "#/components/schemas/Timestamp"},
          "stoppedAt": {"$ref": "#/components/schemas/Timestamp"},
          "updatedAt": {"$ref": "#/components/schemas/Timestamp"}
        }
      },
      "OperationKind": {"type": "string", "enum": ["create", "start", "drain", "stop", "checkpoint", "delete", "cancel_exec", "cancel_terminal"]},
      "OperationState": {"type": "string", "enum": ["pending", "running", "succeeded", "failed", "cancelled"]},
      "Operation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "sandboxId", "kind", "state", "requestId", "createdAt", "updatedAt"],
        "properties": {
          "id": {"$ref": "#/components/schemas/OpaqueID"},
          "sandboxId": {"$ref": "#/components/schemas/OpaqueID"},
          "kind": {"$ref": "#/components/schemas/OperationKind"},
          "state": {"$ref": "#/components/schemas/OperationState"},
          "requestId": {"$ref": "#/components/schemas/CorrelationID"},
          "sandbox": {"$ref": "#/components/schemas/Sandbox"},
          "error": {"$ref": "#/components/schemas/Problem"},
          "createdAt": {"$ref": "#/components/schemas/Timestamp"},
          "startedAt": {"$ref": "#/components/schemas/Timestamp"},
          "completedAt": {"$ref": "#/components/schemas/Timestamp"},
          "updatedAt": {"$ref": "#/components/schemas/Timestamp"}
        }
      },
...
      "ShellCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": ["mode", "command"],
        "properties": {
          "mode": {"const": "shell"},
          "command": {"type": "string", "minLength": 1, "maxLength": 1048576}
        }
      },
      "ArgvCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": ["mode", "executable", "arguments"],
        "properties": {
          "mode": {"const": "argv"},
          "executable": {"type": "string", "minLength": 1, "maxLength": 4096},
          "arguments": {"type": "array", "maxItems": 4096, "items": {"type": "string", "maxLength": 131072}}
        }
      },
      "Command": {
        "oneOf": [
          {"$ref": "#/components/schemas/ShellCommand"},
          {"$ref": "#/components/schemas/ArgvCommand"}
        ],
        "discriminator": {"propertyName": "mode", "mapping": {"shell": "#/components/schemas/ShellCommand", "argv": "#/components/schemas/ArgvCommand"}}
      },
...
      "ExecOutput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["stdoutBase64", "stderrBase64"],
        "properties": {
          "stdoutBase64": {"type": "string", "contentEncoding": "base64"},
          "stderrBase64": {"type": "string", "contentEncoding": "base64"}
        }
      },
      "ExecExited": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "exitCode", "output"],
        "properties": {
          "kind": {"const": "exited"},
          "exitCode": {"type": "integer", "minimum": 0, "maximum": 255},
          "signal": {"type": "integer", "minimum": 1, "maximum": 64},
          "output": {"$ref": "#/components/schemas/ExecOutput"}
        }
      },
      "SpawnFailureKind": {"type": "string", "enum": ["not_found", "permission_denied", "invalid_cwd", "malformed_executable"]},
      "ExecSpawnFailed": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "reason", "message"],
        "properties": {
          "kind": {"const": "spawn_failed"},
          "reason": {"$ref": "#/components/schemas/SpawnFailureKind"},
          "message": {"type": "string", "minLength": 1, "maxLength": 2048}
        }
      },
      "ExecDeadlineExceeded": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "elapsedMilliseconds", "output"],
        "properties": {
          "kind": {"const": "deadline_exceeded"},
          "elapsedMilliseconds": {"type": "integer", "format": "int64", "minimum": 0},
          "output": {"$ref": "#/components/schemas/ExecOutput"}
        }
      },
      "ExecCancelled": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "output"],
        "properties": {
          "kind": {"const": "cancelled"},
          "output": {"$ref": "#/components/schemas/ExecOutput"}
        }
      },
      "ExecOutputExhausted": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "limitBytes", "output"],
        "properties": {
          "kind": {"const": "output_exhausted"},
          "limitBytes": {"type": "integer", "format": "int64", "minimum": 1},
          "output": {"$ref": "#/components/schemas/ExecOutput"}
        }
      },
      "InfrastructureFailureKind": {
        "type": "string",
        "enum": ["transport", "admission", "generation_fenced", "lease_fenced", "guest_agent", "execution_node", "service"]
      },
      "ExecInfrastructureFailed": {
        "type": "object",
        "additionalProperties": false,
        "required": ["kind", "reason", "retryable", "message"],
        "properties": {
          "kind": {"const": "infrastructure_failed"},
          "reason": {"$ref": "#/components/schemas/InfrastructureFailureKind"},
          "retryable": {"type": "boolean"},
          "message": {"type": "string", "minLength": 1, "maxLength": 2048}
        }
      },
      "ExecOutcome": {
        "oneOf": [
          {"$ref": "#/components/schemas/ExecExited"},
          {"$ref": "#/components/schemas/ExecSpawnFailed"},
          {"$ref": "#/components/schemas/ExecDeadlineExceeded"},
          {"$ref": "#/components/schemas/ExecCancelled"},
          {"$ref": "#/components/schemas/ExecOutputExhausted"},
          {"$ref": "#/components/schemas/ExecInfrastructureFailed"}
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "exited": "#/components/schemas/ExecExited",
            "spawn_failed": "#/components/schemas/ExecSpawnFailed",
            "deadline_exceeded": "#/components/schemas/ExecDeadlineExceeded",
            "cancelled": "#/components/schemas/ExecCancelled",
            "output_exhausted": "#/components/schemas/ExecOutputExhausted",
            "infrastructure_failed": "#/components/schemas/ExecInfrastructureFailed"
          }
        }
      },
      "SessionState": {"type": "string", "enum": ["open", "detached", "closing", "closed"]},
...
      "StreamCreditFrame": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "sequence", "bytes"],
        "properties": {
          "type": {"const": "credit"},
          "sequence": {"type": "integer", "format": "int64", "minimum": 0},
          "bytes": {"type": "integer", "format": "int64", "minimum": 1}
        }
      },
...
      "StreamCancelFrame": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "sequence"],
        "properties": {
          "type": {"const": "cancel"},
          "sequence": {"type": "integer", "format": "int64", "minimum": 0}
        }
      },
      "StreamOutcomeFrame": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "sequence", "outcome"],
        "properties": {
          "type": {"const": "outcome"},
          "sequence": {"type": "integer", "format": "int64", "minimum": 0},
          "outcome": {"$ref": "#/components/schemas/ExecOutcome"}
        }
      },
      "CreateTerminalRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["command", "environment", "rows", "columns", "deadlineMilliseconds", "detachable"],
        "properties": {
          "command": {"$ref": "#/components/schemas/Command"},
          "cwd": {"$ref": "#/components/schemas/WorkspacePath"},
          "environment": {"$ref": "#/components/schemas/StringMap"},
          "rows": {"type": "integer", "minimum": 1, "maximum": 1000},
          "columns": {"type": "integer", "minimum": 1, "maximum": 1000},
          "deadlineMilliseconds": {"type": "integer", "format": "int64", "minimum": 1},
          "detachable": {"type": "boolean"}
        }
      },
      "TerminalSession": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "sandboxId", "generation", "state", "websocketUrl", "subprotocol", "nextClientSequence", "expiresAt"],
        "properties": {
          "id": {"$ref": "#/components/schemas/OpaqueID"},
          "sandboxId": {"$ref": "#/components/schemas/OpaqueID"},
          "generation": {"type": "integer", "format": "int64", "minimum": 1},
          "state": {"$ref": "#/components/schemas/SessionState"},
          "websocketUrl": {"type": "string", "format": "uri", "maxLength": 2048},
          "subprotocol": {"const": "secondbox.terminal.v1"},
          "nextClientSequence": {"type": "integer", "format": "int64", "minimum": 0},
          "expiresAt": {"$ref": "#/components/schemas/Timestamp"}
        },
        "x-secondbox-websocket-frames": {"$ref": "#/components/schemas/TerminalFrame"}
      },
      "TerminalFrame": {
        "oneOf": [
          {"$ref": "#/components/schemas/TerminalInputFrame"},
          {"$ref": "#/components/schemas/TerminalOutputFrame"},
          {"$ref": "#/components/schemas/TerminalResizeFrame"},
          {"$ref": "#/components/schemas/StreamCreditFrame"},
          {"$ref": "#/components/schemas/StreamCancelFrame"},
          {"$ref": "#/components/schemas/StreamOutcomeFrame"}
        ],
        "discriminator": {"propertyName": "type"}
      },
      "TerminalInputFrame": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "sequence", "dataBase64"],
        "properties": {
          "type": {"const": "terminal_input"},
          "sequence": {"type": "integer", "format": "int64", "minimum": 0},
          "dataBase64": {"type": "string", "contentEncoding": "base64"}
        }
      },
      "TerminalOutputFrame": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "sequence", "dataBase64"],
        "properties": {
          "type": {"const": "terminal_output"},
          "sequence": {"type": "integer", "format": "int64", "minimum": 0},
          "dataBase64": {"type": "string", "contentEncoding": "base64"}
        }
      },
      "TerminalResizeFrame": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "sequence", "rows", "columns"],
        "properties": {
          "type": {"const": "resize"},
          "sequence": {"type": "integer", "format": "int64", "minimum": 0},
          "rows": {"type": "integer", "minimum": 1, "maximum": 1000},
          "columns": {"type": "integer", "minimum": 1, "maximum": 1000}
        }
      },
...
      "ProblemCode": {
        "type": "string",
        "enum": [
          "invalid_request",
          "authentication_failed",
          "authorization_failed",
          "not_found",
          "idempotency_conflict",
          "precondition_failed",
          "state_conflict",
          "generation_fenced",
          "lease_fenced",
          "profile_unavailable",
          "quota_exceeded",
          "limit_exceeded",
          "guest_unavailable",
          "execution_node_unavailable",
          "dependency_unavailable",
          "internal_error",
          "wait_expired"
        ]
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": ["field", "reason"],
        "properties": {
          "field": {"type": "string", "minLength": 1, "maxLength": 256},
          "reason": {"type": "string", "minLength": 1, "maxLength": 1024}
        }
      },
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "title", "status", "code", "requestId", "retryable"],
        "properties": {
          "type": {"type": "string", "format": "uri", "maxLength": 2048},
          "title": {"type": "string", "minLength": 1, "maxLength": 256},
          "status": {"type": "integer", "minimum": 400, "maximum": 599},
          "code": {"$ref": "#/components/schemas/ProblemCode"},
          "requestId": {"$ref": "#/components/schemas/CorrelationID"},
          "retryable": {"type": "boolean"},
          "details": {"type": "array", "maxItems": 32, "items": {"$ref": "#/components/schemas/ProblemDetail"}}
        }
      }
    }
  }
}